Rename `bank` to `wbank`
