+++ app/app.go
@@ -105,13 +105,13 @@ import (
 	assetkeeper "github.com/CoreumFoundation/coreum/x/asset/keeper"
 	assettypes "github.com/CoreumFoundation/coreum/x/asset/types"
 	"github.com/CoreumFoundation/coreum/x/auth/ante"
-	wbank "github.com/CoreumFoundation/coreum/x/bank"
-	wbankkeeper "github.com/CoreumFoundation/coreum/x/bank/keeper"
 	deterministicgastypes "github.com/CoreumFoundation/coreum/x/deterministicgas/types"
 	"github.com/CoreumFoundation/coreum/x/feemodel"
 	feemodelkeeper "github.com/CoreumFoundation/coreum/x/feemodel/keeper"
 	feemodeltypes "github.com/CoreumFoundation/coreum/x/feemodel/types"
 	wasmtypes "github.com/CoreumFoundation/coreum/x/wasm/types"
+	"github.com/CoreumFoundation/coreum/x/wbank"
+	wbankkeeper "github.com/CoreumFoundation/coreum/x/wbank/keeper"
 	// this line is used by starport scaffolding # stargate/app/moduleImport
 )
