+++ x/wbank/module.go
@@ -1,4 +1,4 @@
-package bank
+package wbank
 
 import (
 	"github.com/cosmos/cosmos-sdk/codec"
@@ -8,7 +8,7 @@ import (
 	bankkeeper "github.com/cosmos/cosmos-sdk/x/bank/keeper"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 
-	"github.com/CoreumFoundation/coreum/x/bank/keeper"
+	"github.com/CoreumFoundation/coreum/x/wbank/keeper"
 )
 
 // AppModuleBasic defines the basic application module used by the wrapped bank module.
